Clippy.

+++ fidget/src/core/vm/alloc.rs
@@ -125,9 +125,9 @@ impl RegisterAllocator {
     /// Memory is treated as unlimited; if we don't have any spare slots, then
     /// we'll assign a new one (incrementing `self.total_slots`).
     ///
-    /// > If there's one thing I love  
-    /// > It's an infinite resource  
-    /// > If there's one thing worth loving  
+    /// > If there's one thing I love
+    /// > It's an infinite resource
+    /// > If there's one thing worth loving
     /// > It's a surplus of supplies
     #[inline]
     fn get_memory(&mut self) -> u32 {
@@ -146,7 +146,7 @@ impl RegisterAllocator {
     /// This is useful when deciding which register to evict to make room
     #[inline]
     fn oldest_reg(&mut self) -> u8 {
-        self.register_lru.pop() as u8
+        self.register_lru.pop()
     }
 
     /// Returns the slot allocated to the given node
